requirements.txt
Flask>=3.0.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
gunicorn>=22.0.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
line-bot-sdk>=3.9.0
PyJWT==2.8.0
cryptography>=42.0.0
google-cloud-firestore>=2.16.0
python-json-logger>=2.0.0
requests>=2.31.0
pandas==2.2.2
openpyxl==3.1.3